move form and error hfunctions into helper files
